Exercicio_Modulo9_TratamentoException_VersaoFinal

+++ threadpool/src/main/java/br/com/mentorama/threadpool/controllers/AlunoController.java
@@ -1,14 +1,16 @@
 package br.com.mentorama.threadpool.controllers;
 
-import br.com.mentorama.threadpool.exceptions.RegistroInexistenteException;
+
 import br.com.mentorama.threadpool.models.Aluno;
 import br.com.mentorama.threadpool.services.AlunoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
 @RequestMapping("/aluno")
@@ -32,8 +34,12 @@ public class AlunoController {
     @GetMapping("/{id}")
     public CompletableFuture<Aluno> findById(@PathVariable ("id") Long id) {
         System.out.println("Controller Thread: " + Thread.currentThread().getName());
-            return this.alunoService.findById(id).thenApply(aluno -> aluno
-                    .orElseThrow(RegistroInexistenteException::new));
+            return this.alunoService.findById(id).thenApply(aluno -> { try {
+                return aluno.get();
+            } catch (NoSuchElementException e) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            }
+        });
     }
 
     @PostMapping
@@ -49,13 +55,12 @@ public class AlunoController {
     }
 
     @DeleteMapping("/{id}")
-    public CompletableFuture<Aluno> delete(@PathVariable ("id") Long id) {
+    public CompletableFuture<Optional<Aluno>> delete(@PathVariable ("id") Long id) {
         System.out.println("Controller Thread: " + Thread.currentThread().getName());
-        return this.alunoService.delete(id);
-    }
-
-    @ExceptionHandler({RegistroInexistenteException.class})
-    public ResponseEntity <String> handle(final RegistroInexistenteException e){
-       return new ResponseEntity<> ("Id Aluno Não Encontrado", HttpStatus.NOT_FOUND);
+         if (this.alunoService.findById(id).join().isPresent()){
+            return this.alunoService.delete(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 }
